backend/requirements.txt
Flask
mysqlclient
pycrypto